refactor: create entity for counterparty
