+++ packages/api/src/hooks/p2p/entity/counterparty/index.ts
@@ -1,0 +1,1 @@
+export * as counterparty from './p2p-advertiser-relations';
