+++ packages/api/src/hooks/p2p/entity/counterparty/p2p-advertiser-relations/index.ts
@@ -1,0 +1,1 @@
+export { default as useUpdate } from './useAdvertiserRelations';
