+++ packages/api/src/hooks/p2p/entity/index.ts
@@ -1,1 +1,2 @@
 export * from './payment-method';
+export * from './counterparty';
